Add Datastore interface to wrap db connection
Using MySQL database driver
