+++ go.mod
@@ -2,4 +2,7 @@ module ether
 
 go 1.13
 
-require github.com/gorilla/mux v1.7.3
+require (
+	github.com/go-sql-driver/mysql v1.5.0
+	github.com/gorilla/mux v1.7.3
+)
